+++ src/app/addplan/page.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import AddPlanForm from '@/components/addPlan/AddPlanForm';
+import AddPlanForm from '@/components/addPlan/AddPlanContents';
 
 export default function AddPlan() {
   return (
